In ggpubr::ggarrange(), I added align = "v".
In this way, the VERTICAL reference lines in the grid areas of the 4 panels will be VERTICALLY aligned together, even when they have y-axis tick labels with different lengths.

This change makes sure that the top-left/right panel and the bottom-left/right panel has the same canvas size.

All the bugs in the previous versions of summary plots are fixed in this commit.